Download specific version of `antlr`
Currently the `generate` target of the _Makefile_ expects the `antlr4`
binary in the path. This means that the developer running it will have
to download it in advance. This isn't very convenient, and it also
introduces the risk of using an incorrect version. To improve that this
patch changes the _Makefile_ so that it will automatically download and
use always the same version.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>